Delete 3 Front matter sections from .md file
We might want to consider deleting the Abstract, Keywords, and Acknowledgements sections from the .md file. In the docx creation these sections can be included in the Front Pages docx file that will be pasted to the front of the entire docx document. Deletion of these sections would affect the generation of the epub, pdf, etc. documents.