Scope section edits
For the quote in the first paragraph, do we have a source for this quote? If not, let's paraphrase it. Possibly "Reference architectures provide credible guidance and boundaries for architecture instantiations and solutions within a subject area."

The @sec:production does not create the correct section reference in the docx document (i.e., it says "... as outlined in sec. 5.3." but Section 5.3 does not contain the description of the three phases of the NBD-PWG work). Also, @sec:production comes out in the docx document as "sec. X.X" but it should be "Section X.X" this happens with the other @sec: references.