List of Specifications edits
The Specification Table should have a title and table number. Do we want to order the specifications within the table to match the order that we discuss them in the rest of the document?

Also, should there be two images (from ![Provider view]... and ![Resource view]...)? Only one image appears in the docx document.