Commenting out incomplete sections
The sections "Compute Management - Containers" and "Compute Management - Functions" only contain the following text "This section is planned for a future version." We are trying to remove references to future work from the volumes since, in theory, version 3 is the last version. Can we comment out these sections? The sections can be added back if content is developed during the public content period.